requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.0
black==26.1.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.104.1
h11==0.16.0
httptools==0.7.1
idna==3.11
mypy_extensions==1.1.0
neo4j==5.15.0
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
platformdirs==4.5.1
pyasn1==0.6.2
pycparser==3.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.6
pytokens==0.4.0
pytz==2025.2
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
